modif: S3 file upload

+++ src/main/java/com/example/alwayswin/service/impl/FigureServiceImpl.java
@@ -1,22 +1,25 @@
 package com.example.alwayswin.service.impl;
 
+import com.amazonaws.AmazonServiceException;
+import com.amazonaws.regions.Regions;
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.AmazonS3ClientBuilder;
+import com.amazonaws.services.s3.transfer.TransferManager;
+import com.amazonaws.services.s3.transfer.TransferManagerBuilder;
+import com.amazonaws.services.s3.transfer.Upload;
 import com.example.alwayswin.config.WebMvcConfig;
 import com.example.alwayswin.entity.Figure;
 import com.example.alwayswin.mapper.FigureMapper;
 import com.example.alwayswin.service.FigureService;
-import com.example.alwayswin.utils.RandomStringUtil;
 import org.apache.commons.beanutils.BeanUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.File;
 import javax.annotation.Resource;
 import java.sql.Timestamp;
-import java.text.SimpleDateFormat;
-import java.util.Date;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -25,6 +28,10 @@ import java.util.Map;
 public class FigureServiceImpl implements FigureService {
     private Logger logger = LoggerFactory.getLogger(getClass());
 
+    private final String BUCKET_NAME = "alwayswin-figures";
+    private final String BUCKET_URL = "https://alwayswin-figures.s3.amazonaws.com/";
+    private final Regions REGION = Regions.US_EAST_1;
+
     @Resource
     private FigureMapper figureMapper;
 
@@ -32,50 +39,6 @@ public class FigureServiceImpl implements FigureService {
 
     public FigureServiceImpl(FigureMapper figureMapper) {
         this.figureMapper = figureMapper;
-    }
-
-    public String upload(MultipartFile file) {
-        String fileName = StringUtils.cleanPath(file.getOriginalFilename());
-
-        try {
-            if (fileName == null) {
-                logger.debug("File not uploaded successfully");
-                return null;
-            }
-            if (fileName.contains("..")) {
-                logger.debug("File has invalid filename");
-                return null;
-            }
-            File dir = new File(uploadPath);
-            if (!dir.exists()){
-                if (dir.mkdir()) {
-                    logger.info(String.format("Create file uploading directory-%s", uploadPath));
-                }else {
-                    logger.warn(String.format("Fail to create file uploading directory-%s", uploadPath));
-                    return null;
-                }
-            }
-
-            SimpleDateFormat simpleDateFormat;
-            simpleDateFormat = new SimpleDateFormat("YYYYMMDD");
-            // Requiring distinctive name -> random generator
-            Date date = new Date();
-            String str = simpleDateFormat.format(date);
-            String imgRan = RandomStringUtil.createRandomString(5);
-            String intervalName = str + imgRan;
-            String displayName = intervalName + "_" + fileName;
-            fileName = uploadPath + displayName;
-            logger.info(String.format("Uploaded file-%s; Directory-%s",displayName , uploadPath));
-
-            File dest = new File(fileName);
-            file.transferTo(dest);
-            logger.info("File " + fileName + " uploaded!");
-
-            return fileName;
-        } catch (Exception e) {
-            logger.warn(e.getMessage(), String.format("[%s] file upload failed", fileName));
-        }
-        return null;
     }
 
     @Override
@@ -130,9 +93,16 @@ public class FigureServiceImpl implements FigureService {
         return figureMapper.add(figure);
     }
 
-    // todo: delete from server as well
     @Override
     public int deleteFigure(int fid) {
+        Figure figure = getFigureByFid(fid);
+        if (figure == null)
+            return 0;
+        String url = figure.getUrl();
+        // 先从bucket里删掉
+        deleteFigureFromS3(url);
+
+        // 再从数据库删掉
         return figureMapper.delete(fid);
     }
 
@@ -144,4 +114,78 @@ public class FigureServiceImpl implements FigureService {
             figureMapper.update(oldThumbnail);
         }
     }
+
+    public String uploadFile(String localPath, String s3FolderName) {
+
+        File f = new File(localPath);
+        TransferManager transferManager = TransferManagerBuilder.standard().build();
+        String s3Filename = null;
+        try {
+            s3Filename = getS3Filename(localPath, s3FolderName);
+            transferManager.upload(BUCKET_NAME, s3Filename, f);
+        } catch (AmazonServiceException e) {
+            logger.error(e.getErrorMessage());
+            return null;
+        }
+        finally {
+            transferManager.shutdownNow();
+        }
+        return BUCKET_URL + s3Filename;
+    }
+
+
+    public List<String> uploadFileList(String[] localPaths, String s3FolderName) {
+        // convert the file paths to a list of File objects (required by the
+        // uploadFileList method)
+        ArrayList<File> files = new ArrayList<>();
+        List<String> urlList = new ArrayList<>();
+        for (String path : localPaths) {
+            files.add(new File(path));
+            urlList.add(BUCKET_URL + getS3Filename(path, s3FolderName));
+        }
+
+        TransferManager transferManager = TransferManagerBuilder.standard().build();
+        try {
+            transferManager.uploadFileList(BUCKET_NAME,
+                    s3FolderName,
+                    new File("."),   // directory
+                    files);
+
+        } catch (AmazonServiceException e) {
+            logger.error(e.getErrorMessage());
+            return null;
+        }
+        finally {
+            transferManager.shutdownNow();
+        }
+        return urlList;
+    }
+
+
+    private String getS3Filename(String localPath, String s3FolderName) {
+        String[] strings = localPath.split("/");
+        String originalFilename = strings[strings.length - 1];  // 最后一个是文件名
+
+        String s3Filename = null;
+        if (s3FolderName != null) {
+            s3Filename = s3FolderName + '/' + originalFilename;
+        } else {
+            s3Filename = originalFilename;
+        }
+        return s3Filename;
+    }
+
+
+    private void deleteFigureFromS3(String url) {
+//        url https://alwayswin-figures.s3.amazonaws.com/product-figure/default-product-thumbnail.png
+        String[] strings = url.substring("https://".length()).split("/", 2);
+        String keyName = strings[1];
+        final AmazonS3 s3 = AmazonS3ClientBuilder.standard().withRegion(REGION).build();
+        try {
+            s3.deleteObject(BUCKET_NAME, keyName);
+        } catch (AmazonServiceException e) {
+            logger.error(e.getErrorMessage());
+        }
+    }
+
 }
